move ClientProperties to its own file
this file was getting a bit crowded
